Rebase/original 3.0.0 pre.8 (#8)
* Improve API documentation

* Improve documentation of VideoStreamTrack

* Improve documentation of MediaStream

* Improve documentation of RTCConfiguration

* Improve documentation of MediaStreamTrack

* Improve documentation of AudioStreamTrack

* Improve documentation of WebRTC

* On the way.

* Improve documentation of RTCRtpTransceiver

* On the way.

* Improve documentation of RTCIceServer

* Refine examples

* On the way.

* Added summary for RTCDataChannelInit class.

* Added Exceptions doc to Send().

* Improve documentation of RTCRtpSender

* Improved the example in RTCDataChannel.

* Improve examples

* Improve examples

* Added some Remarks.

* Improve examples

* Improve examples

* Improve examples

* Improve examples

* Improve examples

* Improve examples

* Improve examples

* Adding Remarks.

* Added Remarks.

* Added sample programs.

* Fixed typo "chennel".

* Imroved description for ordered.

* modifying summaries.

* Improved protocol property.

* Done for RTCDataChannelInit.

* Update Runtime/Scripts/WebRTC.cs


* wip.

* Improve documentation of RTCRtpEncodingParameters

* Filled the all.

* Added more samples.

* Improved exception.

* fixing source code formatting errors.

* refactor: move CameraExtension class to its own file (#1064)

* Update Runtime/Scripts/MediaStreamTrack.cs


* Update Runtime/Scripts/AudioStreamTrack.cs


* Update Runtime/Scripts/AudioStreamTrack.cs


* Update Runtime/Scripts/AudioStreamTrack.cs


* Update Runtime/Scripts/AudioStreamTrack.cs


* Update Runtime/Scripts/WebRTC.cs


* Update Runtime/Scripts/WebRTC.cs


* Update Runtime/Scripts/WebRTC.cs


* Update Runtime/Scripts/WebRTC.cs


* gitignore: ignore UserSettings and Logs (#1068)

* Update Runtime/Scripts/RTCDataChannel.cs


* Update Runtime/Scripts/RTCDataChannel.cs


* Update Runtime/Scripts/RTCDataChannel.cs


* Update Runtime/Scripts/RTCDataChannel.cs


* Update Runtime/Scripts/RTCDataChannel.cs


* Update Runtime/Scripts/RTCDataChannel.cs


* doc: improve the API doc of RTCSessionDescription (#1067)

* initial improvement 
* rewrite 
* add sample code

* project: add a symbolic link to the Documentation folder (#1070)

* doc: improve the API doc of CameraExtension (#1073)

* add documentation
* add class summary
* simplify code
* update example

* readme: update Unity badges (#1074)

* release: 3.0.0-pre.8 (#1075)

* update package.json
* update changelog
* update readme
* update ValidationExceptions.json
* update install.md
* update bug_report.yml

* Feat: libwebrtc update m124 (#1)

* update buildscript on mac

* update buildscript on win

* update buildscript on linux

* update android buildscript

* update windows libwebrtc buid

* update mac build(disable replace task queue)

* update deployment target (#2)

* fix linux plugin build on ubuntu 22.04 (#3)

* change package and readme (#5)

* Fix: createimage vulkan android own (#4)

* add error logs

* add VkMemoryDedicatedAllocateInfo on android

* update sample project settings

* fix project

* update plugin win/linux/android

* update plugin mac/ios

* Fix: crash android vulkan (#7)

* fix meta

* workaround avoid crash access to i420

* update android aar

* update version 3.0.2-pre.7

---------
