+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "com.kannan-xiao4.webrtc",
   "displayName": "WebRTC",
-  "version": "3.0.2-pre.7",
+  "version": "3.0.0-pre.8",
   "unity": "2022.3",
   "description": "The WebRTC package provides browsers and mobile applications with Real-Time Communications (RTC) capabilities.",
   "keywords": [
